requirements.txt
faiss-cpu~=1.11.0.post1
numpy~=2.3.1
sentence-transformers~=5.0.0
pandas~=2.3.1
joblib~=1.5.1
requests~=2.32.4